Assignment review updates
This should change the bug where even guessing correctly in the final chance still prints a "losing" message. The bug where inputting an invalid guess the application terminates and throws an error never happened on my end and still does not so I'm not exactly sure how or why that happened when it was reviewed but if it continues I will look more deeply.